Refactor directory tests to separate file
